test(flow-filter): Move lib.rs' tests to dedicated file
The tests have grown quite a lot in flow-filter/src/lib.rs and it is not
very convenient to have such a large portion of the file occupied by
tests. Let's move them to a separate file.

Submodules retain their own tests.
